ref: Clean up jspsych.js as index,js
