--- .codesandbox/workspace.json
@@ -1,20 +1,0 @@
-{
-  "responsive-preview": {
-    "Mobile": [
-      320,
-      675
-    ],
-    "Tablet": [
-      1024,
-      765
-    ],
-    "Desktop": [
-      1400,
-      800
-    ],
-    "Desktop  HD": [
-      1920,
-      1080
-    ]
-  }
-}
